Reverse Proxy & X-Forwarded-* Headers
Now that we're behind a load balancer... we have a new, minor, but important problem. Suppose that, in our app, we want to get the user's IP address. So, $ request->getClientIp(). Guess what? That will now be the wrong IP address! Instead of being the IP of the user, it will always be the IP of the load balancer!

In fact, a bunch of things will be wrong. For example, $ request->isSecure() will return false, even if the user is accessing our site over https. The port and even the host might be wrong!

This is a classic problem when you're behind a proxy, like a load balancer. When the load balancer sends the request back to our server, it changes a few things: the REMOTE_ADDR header is changed to be the load balancer's IP address. And if the original request was an https connection on port 443, the new request will appear insecure on port 80. That's because the load balancer handled the SSL stuff.

To help us, the load balancer sets the original information on a few headers: X-Forwarded-For holds the original IP address and X-Forwarded-Proto will be set to http or https.

Tip

There are some standards, but the exact headers used can vary from proxy to proxy.

This means that our app needs to be smart enough to read these headers, instead of the normal ones. Symfony doesn't do this automatically, because it could be a security risk. You need to configure it explicitly.